Fix Combat log dashboards
Fix Combat log dashboards

+++ src/CombatParser/CombatParser.Application/Queries/Dashboards/GetHealSpells/GetHealSpellsHandler.cs
@@ -3,11 +3,11 @@ using MediatR;
 
 namespace CombatParser.Application.Queries.Dashboards.GetHealSpells;
 
-internal class GetHealSpellsHandler(IDashboardRepository repository) : IRequestHandler<GetHealSpellsQuery, Dictionary<string, int>>
+internal class GetHealSpellsHandler(IDashboardRepository repository) : IRequestHandler<GetHealSpellsQuery, Dictionary<string, long>>
 {
     private readonly IDashboardRepository _repository = repository;
 
-    public async Task<Dictionary<string, int>> Handle(GetHealSpellsQuery request, CancellationToken cancellationToken)
+    public async Task<Dictionary<string, long>> Handle(GetHealSpellsQuery request, CancellationToken cancellationToken)
     {
         var spells = await _repository.GetHealSpellsAsync(request.CombatLogId, cancellationToken);
