+++ src/CombatParser/CombatParser.Application/Queries/Dashboards/GetHealSpells/GetHealSpellsQuery.cs
@@ -4,4 +4,4 @@ namespace CombatParser.Application.Queries.Dashboards.GetHealSpells;
 
 public record GetHealSpellsQuery(
     int CombatLogId
-    ) : IRequest<Dictionary<string, int>>;
+    ) : IRequest<Dictionary<string, long>>;
